Moved CRUD to Drivers\Database Added GuzzleHttp as a Cobalt dependency
