+++ classes/Drivers/Database.php
@@ -1,31 +1,28 @@
 <?php
 
-namespace CRUD;
+namespace Drivers;
 
-use Exception;
 use MongoDB\BSON\ObjectId;
 use MongoDB\BSON\UTCDateTime;
-use CRUD\Exceptions\ValidationFailed;
+use Validation\Exceptions\ValidationFailed;
 
-abstract class CRUD {
+abstract class Database {
     public $db = __APP_SETTINGS__['database'];
 
-    /**
-     * @return string the name of the database collection (table)
-     */
-    abstract function set_collection();
+    /** @return string the name of the database collection (table) */
+    abstract function get_collection_name();
 
     function __construct() {
-        $this->collection = db_cursor($this->set_collection(), $this->db);
+        $this->collection = db_cursor($this->get_collection_name(), $this->db);
     }
 
-    /** HELPERS */
-    function __id($id = null) {
+    /* HELPERS */
+    final function __id($id = null) {
         if ($id === null) return new ObjectId();
         return new ObjectId($id);
     }
 
-    function __date($date = null) {
+    final function __date($date = null) {
         if ($date === null) return new UTCDateTime();
         switch (gettype($date)) {
             case "string":
@@ -42,48 +39,48 @@ abstract class CRUD {
         throw new ValidationFailed("Invalid date parameter");
     }
 
-    private function date_string_parse($date) {
+    final function date_string_parse($date) {
         $timestamp = strtotime($date);
         if ($timestamp) return new UTCDateTime($date);
         throw new ValidationFailed("Invalid date parameter");
     }
 
-    /** CREATE */
-    function insertOne($document, array $options = []) {
+    /* CREATE */
+    final function insertOne($document, array $options = []) {
         return $this->collection->insertOne($document, $options);
     }
 
-    function insertMany($documents, array $options = []) {
+    final function insertMany($documents, array $options = []) {
         return $this->collection->insertMany($documents, $options);
     }
 
 
-    /** READ */
-    function findOne($filter, array $options = []) {
+    /* READ */
+    final function findOne($filter, array $options = []) {
         return $this->collection->findOne($filter, $options);
     }
 
-    function find($filter = [], array $options = []) {
+    final function find($filter = [], array $options = []) {
         return $this->collection->find($filter, $options);
     }
 
 
-    /** UPDATE */
-    function updateOne($filter, $fields, array $options = []) {
+    /* UPDATE */
+    final function updateOne($filter, $fields, array $options = []) {
         return $this->collection->updateOne($filter, $fields, $options);
     }
 
-    function updateMany($filter, $fields, array $options = []) {
+    final function updateMany($filter, $fields, array $options = []) {
         return $this->collection->updateMany($filter, $fields, $options);
     }
 
 
-    /** DESTROY */
-    function deleteOne($filter, array $options = []) {
+    /* DESTROY */
+    final function deleteOne($filter, array $options = []) {
         return $this->collection->deleteOne($filter, $options);
     }
 
-    function deleteMany($filter, array $options = []) {
+    final function deleteMany($filter, array $options = []) {
         return $this->collection->deleteMany($filter, $options);
     }
 }
