+++ composer.json
@@ -4,6 +4,7 @@
         "erusev/parsedown": "^1.7",
         "matthiasmullie/minify": "^1.3",
         "phpmailer/phpmailer": "^6.4",
-        "stripe/stripe-php": "^7.78"
+        "stripe/stripe-php": "^7.78",
+        "guzzlehttp/guzzle": "^7.3"
     }
 }
